+++ src/plugins/dev-mode/error-messages.ts
@@ -163,7 +163,7 @@ export const ERROR_MESSAGES = {
     RC_STREAM: 'RxReplication pull stream$ threw an error - see .errors for more details',
     RC_PUSH: 'RxReplication push handler threw an error - see .errors for more details',
     RC_PUSH_NO_AR: 'RxReplication push handler did not return an array with the conflicts',
-    RC_P2P_PEER: 'RxReplication P2P Peer has error',
+    RC_WEBRTC_PEER: 'RxReplication WebRTC Peer has error',
     RC_COUCHDB_1: 'replicateCouchDB() url must end with a slash like \'https://example.com/mydatabase/\'',
     RC_COUCHDB_2: 'replicateCouchDB() did not get valid result with rows.',