+++ test/unit.test.ts
@@ -58,7 +58,7 @@ import './unit/replication.test';
 import './unit/replication-graphql.test';
 import './unit/replication-couchdb.test';
 import './unit/replication-websocket.test';
-import './unit/replication-p2p.test';
+import './unit/replication-webrtc.test';
 import './unit/crdt.test';
 import './unit/population.test';
 import './unit/leader-election.test';